+++ math/vector.h
@@ -16,22 +16,22 @@ namespace math {
     static const Type kZero;
     static const Type kEpsilon;
 
-    Vector(int size);
-    Vector(int size, const Type *components);
+    Vector(size_t size);
+    Vector(size_t size, const Type *components);
     Vector(const Vector &other);
 
-    int Size(void) const;
+    size_t Size(void) const;
     Type Distance(const Vector &other) const;
 
     Vector& operator = (const Vector &other);
-    Type& operator [] (int index);
-    const Type& operator [] (int index) const;
+    Type& operator [] (size_t index);
+    const Type& operator [] (size_t index) const;
 
     friend std::istream& operator >> (std::istream &is, Vector &v);
     friend std::ostream& operator << (std::ostream &os, Vector &v);
 
   private:
-    const int size_;
+    const size_t size_;
     std::valarray<Type> components_;
 
     static bool EQ(Type u, Type v);
